Se agrega atributo activated en clases de Ventas y Productos para eliminar dichas clases de forma logica y se modifican servicios de lista para solo retornar registros activos.

+++ src/main/java/tobias/web/rest/ProductoResource.java
@@ -174,6 +174,11 @@ public class ProductoResource {
     @DeleteMapping("/productos/{id}")
     public ResponseEntity<Void> deleteProducto(@PathVariable Long id) {
         log.debug("REST request to delete Producto : {}", id);
+
+        if (!productoRepository.existsById(id)) {
+            throw new BadRequestAlertException("Entity not found", ENTITY_NAME, "idnotfound");
+        }
+
         productoService.delete(id);
         return ResponseEntity
             .noContent()
